requirements_dev.txt
# requirements-dev.txt
-r requirements.txt

# Development/Testing Tools
django-debug-toolbar==4.2.0
django-extensions==3.2.3
pytest-django==4.7.0
coverage==7.6.0  # Updated for Python 3.13
black==23.11.0
flake8==6.1.0
isort==5.12.0